use blocks highlight
