+++ apps/web/app/components/blocks/structure/FooterContainer/FooterContainerForm.vue
@@ -86,7 +86,6 @@ import type { Block } from '@plentymarkets/shop-api';
 
 const { toggleBlockVisibility } = useBlocksVisibility();
 const { footer } = useBlocks();
-const { scrollToBlock, highlightTimeoutToken, highlightedUuid } = useTableOfContents();
 
 const { setEditTitle, clearEditTitle } = useBlockEditTitle();
 
@@ -97,7 +96,6 @@ const colorsOpen = ref(true);
 const editingBlockIndex = ref<number | undefined>(undefined);
 const editingBlockName = ref<string | undefined>(undefined);
 const blockLabels = ref<string[]>([]);
-const currentActiveBlockIndex = ref<number>(-1);
 
 const footerContainer = computed(() => (footer.value ?? {}) as FooterContainerBlock);
 
@@ -152,32 +150,16 @@ const blocks = computed({
   },
 });
 
+const { currentActiveBlockIndex, selectBlock, highlightActiveBlock } = useBlocksHighlight(blocks);
+
 const resolveBlockLabels = async () => {
   blockLabels.value = await Promise.all(blocks.value.map((block) => getBlockDisplayName(block.name)));
-};
-
-const hasValidActiveBlock = computed(() => 
-  currentActiveBlockIndex.value >= 0 && currentActiveBlockIndex.value < blocks.value.length
-);
-
-const selectBlock = (index: number) => {
-  currentActiveBlockIndex.value = index;
-  const block = blocks.value[index];
-  if (block) {
-    scrollToBlock(block.meta.uuid);
-    highlightTimeoutToken.value++;
-  }
 };
 
 const editBlock = (index: number) => {
   editingBlockIndex.value = index;
   editingBlockName.value = footerContainer.value?.content?.[index]?.name;
-  currentActiveBlockIndex.value = index;
-  const block = blocks.value[index];
-  if (block) {
-    scrollToBlock(block.meta.uuid);
-    highlightTimeoutToken.value++;
-  }
+  selectBlock(index);
   setEditTitle(blockLabels.value[index]!);
 };
 
@@ -230,14 +212,7 @@ const deleteBlock = async (index: number) => {
 
 const updateBlocks = (newBlocks: Block[]) => {
   blocks.value = newBlocks;
-  
-  if (hasValidActiveBlock.value) {
-    const activeBlock = blocks.value[currentActiveBlockIndex.value];
-    if (activeBlock) {
-      scrollToBlock(activeBlock.meta.uuid);
-      highlightTimeoutToken.value++;
-    }
-  }
+  highlightActiveBlock();
 };
 
 const toggleBlockVisibilityHandler = (index: number) => {
@@ -256,27 +231,13 @@ const toggleBlockVisibilityHandler = (index: number) => {
   blocks.value = updatedBlocks;
 };
 
-const previousBlocksLength = ref(blocks.value.length);
-
 watch(
   () => blocks.value.map((block) => block.meta.uuid),
-  (newUuids) => {
+  () => {
     resolveBlockLabels();
-    
-    if (newUuids.length > previousBlocksLength.value) {
-      const newBlockIndex = blocks.value.length - 1;
-      selectBlock(newBlockIndex);
-    }
-    
-    previousBlocksLength.value = newUuids.length;
   },
   { immediate: true },
 );
-
-onBeforeUnmount(() => {
-  highlightedUuid.value = '';
-  currentActiveBlockIndex.value = -1;
-});
 
 defineExpose({
   exitEditMode,
