+++ apps/web/app/composables/useBlocksHighlight/index.ts
@@ -1,0 +1,1 @@
+export { useBlocksHighlight } from './useBlocksHighlight';
